+++ CHANGELOG.md
@@ -265,4 +265,5 @@ of enums in the Pump. Add qualifiers to position_def.json for ufid and deptid to
 code format improvements. pump updates to pass query_parms to vivo_query.  Fix bug in uf_examples/manage_columns_filter.
 Improve uf_examples/README.md. New improve_jobcode_description passes pylint
 * **2015-09-03** New improve_title.  vivopump.py now passes pylint.  Move all TODO items to GitHub issues.
-* **2015-09-04** uf_examples/sponsors now demonstrates harvested_by and date_harvested.
+* **2015-09-04** uf_examples/sponsors now demonstrates harvested_by and date_harvested. Updating test cases.
+Continue to get query parameters passed to vivo_query.  New_uri now has parms as an argument for passing to vivo_query